label dtype to float

+++ Code/projs/recsys/Dataset.py
@@ -72,7 +72,7 @@ class d2lCTRDataset(torch.utils.data.Dataset):
     def __init__(self, data_path):
         parser = CategDataParser(data_path, num_cols=[1, 34], set_label_mappers={1:{'0':0, '1':1}})
         dataset = torch.tensor(parser.mapped_data, dtype=torch.int64) # (sample_size, 35)
-        self.label = dataset[:, 0] # (sample_size, )
+        self.label = dataset[:, 0].type(torch.float32) # (sample_size, )
         self.features = dataset[:, 1:] # (sample_size, 34)
         self._num_classes = torch.tensor(parser.feat_dims, dtype=torch.int64) # (34,)
